requirements.txt
asgiref==3.2.10
coverage==5.2.1
Django==3.1
djangorestframework==3.11.1
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
